fix(core): 修复发送字节数统计类型错误
- 将 sentBytes 统计中的类型转换移除，避免类型截断问题
- 将 AbstractSynRunnable 中 isCanceled 字段改为 volatile，保证多线程可见性
- 调整 AbstractSynRunnable 中 ReentrantLock 的声明顺序以提高代码清晰度

+++ mica-net-core/src/main/java/net/dreamlu/mica/net/core/WriteCompletionHandler.java
@@ -341,7 +341,7 @@ public class WriteCompletionHandler implements CompletionHandler<Long, WriteComp
 				channelContext.stat.latestTimeOfSentPacket = System.currentTimeMillis();
 				if (tioConfig.statOn) {
 					tioConfig.groupStat.sentBytes.add(bytesWritten);
-					channelContext.stat.sentBytes.addAndGet((int) bytesWritten);
+					channelContext.stat.sentBytes.addAndGet(bytesWritten);
 				}
 			}
